+++ src/components/Forms/FormInput/FormInput.tsx
@@ -1,21 +1,19 @@
 import { InputLine } from "../..";
 
-import { Props } from "./types";
 import s from "../Form.module.css";
+import { Props } from "./types";
 
-function FormInputLine({ tag, label, register, errors, isSubmitting }: Props) {
-  const { ref, name, onBlur, onChange } = register;
+const FormInput = ({ tag, label, register, errors }: Props) => {
   return (
     <>
       <InputLine
-        {...{ ref, name, onBlur, onChange }}
         label={label}
         invalid={errors[tag] ? "true" : "false"}
-        disabled={isSubmitting}
+        {...register(tag)}
       />
       {errors[tag] && <div className={s.formError}>{errors[tag]?.message}</div>}
     </>
   );
-}
+};
 
-export default FormInputLine;
+export default FormInput;
